% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_to_seurat.R
\name{as.tbl_intervalcustom}
\alias{as.tbl_intervalcustom}
\title{Convert GTF to tibble}
\usage{
as.tbl_intervalcustom(x)
}
\arguments{
\item{x}{GTF from import}
}
\value{
data.frame
}
\description{
Convert GTF to tibble
}
