% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_priming_region.R
\name{bin_effectsize}
\alias{bin_effectsize}
\title{calculation mean diff and effect size of utr3 priming}
\usage{
bin_effectsize(df, target_loc = "utr3", col1 = NA, col2 = "id", ctrl = NA)
}
\arguments{
\item{df}{data.frame with count bins}

\item{target_loc}{string for target region, default utr3}

\item{col1}{column designating broader grouping if needed}

\item{col2}{column designating grouping, default to "id" from countlist_to_bins}

\item{ctrl}{target control set in col}
}
\value{
data.frame with stats (mean diff and effecti size)
}
\description{
calculation mean diff and effect size of utr3 priming
}
