% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_button.R
\name{click_button}
\alias{click_button}
\title{Click a Button on a Web Page using RSelenium}
\usage{
click_button(remDr, button_id = NULL, button_class = NULL)
}
\arguments{
\item{remDr}{An object of RSelenium remote driver.}

\item{button_id}{A string specifying the ID of the button to be clicked. If NULL, button_class is used.}

\item{button_class}{A string or vector of strings specifying the class(es) of the button to be clicked.
If NULL, button_id is used.}
}
\value{
Returns TRUE if the button click leads to a change in page length, indicating successful interaction.
Returns FALSE otherwise, or if both button_id and button_class are NULL, or if an error occurs.
}
\description{
This function attempts to click a button on a web page identified either by its ID or class name using RSelenium.
It checks if the page length changes after the button click, indicating a successful interaction.
}
\examples{
\dontrun{
# Assuming 'remDr' is a valid RSelenium remote driver object
click_button(remDr, button_id = "showMoreBtn")
click_button(remDr, button_class = "load-more-button")
}

}
