% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_priming_region.R
\name{cohens_d}
\alias{cohens_d}
\title{calculation cohens_d effect size}
\usage{
cohens_d(x, y)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\value{
cohens_d effect size
}
\description{
calculation cohens_d effect size
}
