% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Collect motorcycle prices with scrapr::zoom()}
\usage{
collect()
}
\value{
A data frame of five rows (types), 4 columns: 
type of motocycle \code{cycle}, \code{price},
the url \code{site}, and \code{date}.
}
\description{
\code{zoom} works with \code{collect} to retrieve motorcycle prices
}
