% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab+collect_byID.R
\name{collect_byID}
\alias{collect_byID}
\title{Retrieve Data for 884 Items from Walmart API}
\usage{
collect_byID()
}
\value{
Returns 884 by 8 data frame, including the six columns \code{grab_byID} returns,
plus the date of retrieval and the item name.
}
\description{
\code{collect_byID} retrieves data for 884 preselected items from the Walmart API.
}
\details{
The items were selected based on 'relevant' results for approximately 170 Entry Level
Items that Walmart sells. The initial 'search' queried by item name rather than ID.
The 884 item IDs and names are included as package data, named \code{dat_itemIDs} and \code{ELI} respectively.
}
