% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_priming_region.R
\name{countlist_to_bins}
\alias{countlist_to_bins}
\title{calculations from countlist, ready to plot}
\usage{
countlist_to_bins(countlist, vec, genes = NA, k = NA)
}
\arguments{
\item{countlist}{output from scraps_to_region_counts and filter_countlist}

\item{vec}{vector of cluster identities}

\item{genes}{if provided, only calculates based on these genes}

\item{k}{quantiles to use, binned by 2000 genes if not given}
}
\value{
data.frame with count bins
}
\description{
calculations from countlist, ready to plot
}
