% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{920 x 8: Walmart API search by Name}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 920 rows and 8 columns.}
\usage{
dat
}
\description{
920 x 8: Walmart API search by Name
}
\keyword{datasets}
