% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datr}
\alias{datr}
\title{884 x 8: Walmart API search by Item ID}
\format{An object of class \code{data.frame} with 884 rows and 8 columns.}
\usage{
datr
}
\description{
884 x 8: Walmart API search by Item ID
}
\keyword{datasets}
