% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_motif.R
\name{DE_PA_motifs}
\alias{DE_PA_motifs}
\title{find motif differences from all differential expression on polyA sites}
\usage{
DE_PA_motifs(motifs, res)
}
\arguments{
\item{motifs}{motif file}

\item{res}{DEXseq results data.frame}
}
\value{
motif table
}
\description{
find motif differences from all differential expression on polyA sites
}
