% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_priming_region.R
\name{dfnot_avg}
\alias{dfnot_avg}
\title{filter region count list and convert to matrices}
\usage{
dfnot_avg(df, dfgene, vec, k, utr5, utr3, cds, intron)
}
\arguments{
\item{df}{one df from output from scraps_to_region_counts}

\item{dfgene}{ranked gene data.frame}

\item{vec}{vector of cluster identities}

\item{k}{number of bins}

\item{utr5}{counts for region}

\item{utr3}{counts for region}

\item{cds}{counts for region}

\item{intron}{counts for region}
}
\value{
percentage binned averages for specific region
}
\description{
filter region count list and convert to matrices
}
