% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ELI}
\alias{ELI}
\title{184 Entry Level Items, by name}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 184 rows and 1 columns.}
\usage{
ELI
}
\description{
184 Entry Level Items, by name
}
\keyword{datasets}
