% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_and_format_dates_DR}
\alias{extract_and_format_dates_DR}
\title{Extract and Format Date from News Element (DR)}
\usage{
extract_and_format_dates_DR(list_of_elements)
}
\arguments{
\item{list_of_elements}{The news article element from which the date is to be extracted.}
}
\value{
A character string representing the formatted date, or NA if the date cannot be determined.
}
\description{
This function extracts the publication date from a news article element
and formats it into a standardized format. It handles different structures
of the date element within the news article.
}
\examples{
\dontrun{
  # Assuming 'list_of_elements' is a list HTML node from a news website
  extract_and_format_dates_DR <- extract_and_format_date(list_of_elements)
}

}
