% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_functions.R
\name{extract_data}
\alias{extract_data}
\title{Extract data from PDF}
\usage{
extract_data(file_name = "figure1.pdf")
}
\description{
This function simulates data
}
\examples{
extract_data()
}
\keyword{simulation}
