% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_functions.R
\name{extract_PDF_data}
\alias{extract_PDF_data}
\title{Extract data from PDF}
\usage{
extract_PDF_data(file_name = "figure1.pdf", integer_values = F,
  x_log_scale = F, y_log_scale = F)
}
\arguments{
\item{file_name}{File to load}

\item{integer_values}{TRUE if the points are expected to be integer values}

\item{x_log_scale}{TRUE if x axis is on a log scale}

\item{y_log_scale}{TRUE if y axis is on a log scale}
}
\description{
This function uses the alignment tools to extract the raw data
}
\examples{
extract_PDF_data()
}
