% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_priming_region.R
\name{filter_countlist}
\alias{filter_countlist}
\title{filter region count list and convert to matrices}
\usage{
filter_countlist(countlist, n_min = 100)
}
\arguments{
\item{countlist}{output from scraps_to_region_counts}

\item{n_min}{minimum number of observations for a gene to be kept}
}
\value{
count list for regions
}
\description{
filter region count list and convert to matrices
}
