% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_date_from_time}
\alias{format_date_from_time}
\title{Format Date from Raw Date String}
\usage{
format_date_from_time(raw_date)
}
\arguments{
\item{raw_date}{A character string representing the raw date to be formatted.}
}
\value{
A character string of the date in 'YYYY-MM-DD' format.
If the input is not a recognized date format or is empty, returns `NA`.
}
\description{
This function converts a raw date string into a standardized date format (YYYY-MM-DD).
It handles various formats, including relative dates like "I dag" (today) or "I går" (yesterday) in Danish,
as well as specific dates with Danish month abbreviations.
}
\examples{
\dontrun{
format_date_from_time("I dag") # returns current date in 'YYYY-MM-DD'
format_date_from_time("I går") # returns yesterday's date in 'YYYY-MM-DD'
format_date_from_time("5. jun") # returns 'current year-06-05'
format_date_from_time("Invalid Date") # returns NA
}

}
