% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_news_elements_DR}
\alias{get_news_elements_DR}
\title{Extract News Elements from a Web Page Source (DR)}
\usage{
get_news_elements_DR(page_source = NULL)
}
\arguments{
\item{page_source}{A character string or xml2 object representing the HTML source of a web page.}
}
\value{
An object of class `xml_nodeset` containing the extracted news elements.
If the page source is NULL or the outlet is not recognized, the function stops with an error message.
}
\description{
This function extracts news-related HTML elements from the provided web page source,
specifically targeting Danish news outlets 'dr' and 'tv2'.
It uses XPath to identify relevant elements within the HTML structure.
}
\examples{
\dontrun{
# Assuming valid HTML content in 'html_content' variable
get_news_elements_DR(page_source = html_content)
}

}
