% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab+collect_byID.R
\name{grab_byID}
\alias{grab_byID}
\title{Retrieve item information from Walmart API}
\usage{
grab_byID(IDs)
}
\arguments{
\item{IDs}{The Walmart ID(s) of item(s) to collect information on.}
}
\value{
A data frame of i rows, where i is the number of IDs provided;
and six basically self explanatory columns: \code{itemId}, \code{name},
\code{salePrice}, \code{categoryPath}, \code{productUrl}, \code{offerType}.
}
\description{
\code{grab_byID} retrieves item information from the Walmart API, by Item ID, and trims
the results to six helpful columns.
}
