% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{grab_date_from_url}
\alias{grab_date_from_url}
\title{Extract Date from URL}
\usage{
grab_date_from_url(url)
}
\arguments{
\item{url}{A character string representing the URL to be parsed.}
}
\value{
A character string of the date in 'YYYY-MM-DD' format if a valid date is found in the URL.
Returns `NA` if no valid date format is detected.
}
\description{
This function parses a given URL to extract a date in the format YYYY-MM-DD.
It is designed to work with URLs that contain dates embedded within their paths.
}
\examples{
\dontrun{
grab_date_from_url("https://example.com/2023-01-01/article-title")
# returns "2023-01-01"

grab_date_from_url("https://example.com/article-title")
# returns NA
}

}
