% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{html_parse}
\alias{html_parse}
\title{Creates a list-column tibble for each function and each file.}
\usage{
html_parse(files, funs = dplyr::funs(tab = rvest::html_table(., fill = TRUE)),
  fail = FALSE, progress = TRUE, parallel = FALSE, save_rds = TRUE,
  overwrite = TRUE, rds_dir = paste0(dirname(files), "_rds"))
}
\arguments{
\item{files}{file paths}

\item{funs}{list of functions to pass. See \code{\link[dplyr]{funs}} for details.}

\item{fail}{should it fail if encounter one single error?}

\item{progress}{should it show progress bar or informative messages?}

\item{parallel}{should it run in parallel?}

\item{save_rds}{should it save a .rds file for each parsed document?}

\item{overwrite}{should it overwrite the rds file?}

\item{rds_dir}{directory where the .rds files will be saved.}
}
\description{
Creates a list-column tibble for each function and each file.
}
