% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_viewstate}
\alias{html_viewstate}
\alias{html_eventval}
\alias{html_tag_value}
\title{Gets tag value by tag id}
\usage{
html_viewstate(html)

html_eventval(html)

html_tag_value(html, id)
}
\arguments{
\item{html}{\code{xml_document} or \code{xml_node} object.}

\item{id}{CSS path or id name}
}
\value{
atomic character
}
\description{
Gets tag value by tag id.
}
