% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-view.R
\name{html_view}
\alias{html_view}
\alias{html_view.response}
\alias{html_view.character}
\title{Visualise response}
\usage{
html_view(response)

\method{html_view}{response}(response)

\method{html_view}{character}(response)
}
\arguments{
\item{response}{response object (from \code{httr}), file path or character
string containing HTML.}
}
\value{
Returns \code{NULL}.
}
\description{
Visualise response in RStudio Viewer (if available) or users browser.
}
