% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_functions.R
\name{load_data}
\alias{load_data}
\title{Load PDF data from figure}
\usage{
load_data(file_name = "figure1.pdf", file_path = "data/figure1.pdf")
}
\description{
This function simulates data
}
\examples{
simulate_data()
}
\keyword{simulation}
