% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_motif.R
\name{motifs_prep}
\alias{motifs_prep}
\title{prepare motif index from oRNAment files}
\usage{
motifs_prep(
 
    rna = "http://rnabiology.ircm.qc.ca/BIF/oRNAment/static/Homo_sapiens_cDNA_oRNAment.csv.gz",
 
    id1 = "http://rnabiology.ircm.qc.ca/BIF/oRNAment/static/Homo_sapiens_string_to_int_ID_conversion.csv.gz",
  id2 = "http://rnabiology.ircm.qc.ca/BIF/oRNAment/static/RBP_id_encoding.csv.gz",
  min_score = 0.5,
  out = "hs_motifs_05"
)
}
\arguments{
\item{rna}{cDNA_oRNAment.csv.gz file}

\item{id1}{string_to_int_ID_conversion.csv.gz file}

\item{id2}{RBP_id_encoding.csv.gz file}

\item{min_score}{filtering score threshold, default at 0.8}

\item{out}{saved file name}
}
\value{
motif idx table
}
\description{
prepare motif index from oRNAment files
}
