% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_test.R
\name{PA_DEXSeq}
\alias{PA_DEXSeq}
\title{Run DEXSeq differential expression on polyA sites}
\usage{
PA_DEXSeq(
  mat,
  cell_ids1,
  cell_ids2,
  fc_cut = 0.5,
  padj_cut = 0.1,
  pseudo_n = 6,
  seed = 1,
  type = c("3'UTR"),
  indep_fil = TRUE
)
}
\arguments{
\item{mat}{PA site count matrix}

\item{cell_ids1}{cell barcodes for group 1}

\item{cell_ids2}{cell barcodes for group 2}

\item{fc_cut}{min fold change cutoff}

\item{padj_cut}{max adjusted p value cutoff}

\item{pseudo_n}{number of pseudobulk profiles to make for each group}

\item{indep_fil}{independent filtering for DEXSeq}

\item{types}{filter to only these types of PA sites, set to NULL to use all}
}
\value{
DEXSeq result table
}
\description{
Run DEXSeq differential expression on polyA sites
}
