% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbp_motif.R
\name{PA_motifs}
\alias{PA_motifs}
\title{find motif differences from 1 pair polyA sites}
\usage{
PA_motifs(site1 = "chr14_77512017", site2 = "chr14_77505997", motifs)
}
\arguments{
\item{site1}{chr and site name for 1}

\item{site2}{chr and site name for 2}

\item{motifs}{motif file}
}
\value{
motif vector
}
\description{
find motif differences from 1 pair polyA sites
}
