% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_to_seurat.R
\name{parse_gtf}
\alias{parse_gtf}
\title{Parse GTF annotation file}
\usage{
parse_gtf(file)
}
\arguments{
\item{file}{SAF file used with Scraps}
}
\value{
data.frame
}
\description{
Parse GTF annotation file
}
