% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_to_seurat.R
\name{parse_saf}
\alias{parse_saf}
\title{Read SAF annotation file}
\usage{
parse_saf(file, types = FALSE, sep = ";")
}
\arguments{
\item{file}{SAF file used with Scraps}

\item{types}{filter to only these types of PA sites, set to NULL to use all}

\item{sep}{separator between fields for full name}
}
\value{
data.frame
}
\description{
Read SAF annotation file
}
