% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_to_seurat.R
\name{parse_saf_pf}
\alias{parse_saf_pf}
\title{Parse SAF annotation file for position factor (pf) values}
\usage{
parse_saf_pf(file, types = "3'UTR", alt_only = TRUE)
}
\arguments{
\item{file}{SAF file used with Scraps}

\item{types}{filter to only these types of PA sites, set to NULL to use all}

\item{alt_only}{if TRUE, only keep genes with alternative polyA sites}
}
\value{
data.frame with SAF first column and position factor by gene
}
\description{
Parse SAF annotation file for position factor (pf) values
}
