% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{process_news_elements_DR}
\alias{process_news_elements_DR}
\title{Process News Elements from DR into Structured List}
\usage{
process_news_elements_DR(html_elements)
}
\arguments{
\item{html_elements}{A list of news article elements to be processed.}
}
\value{
A list of lists, where each inner list contains the URL, headline,
        topic, and date of a news article.
}
\description{
Processes a list of news elements, extracting key details such as the URL,
headline, topic, and publication date. It utilizes the 'extract_and_format_date'
function to format the date.
}
\examples{
\dontrun{
  # Assuming 'html_elements' is a list of HTML nodes from a news website
  process_news_elements_DR <- process_news_elements(html_elements)
}
}
