% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_page_for_button.R
\name{query_page_for_button}
\alias{query_page_for_button}
\title{Query page for button}
\usage{
query_page_for_button(remDr, query_pattern = NULL, query_attr = NULL)
}
\arguments{
\item{remDr}{A remote driver object see remoteDr}

\item{query_pattern}{A string or vector of strings specifying words used to query class(es) of elements with button tags}

\item{query_attr}{A string specifying the attribute to search for. Options are 'class' or 'id'.}
}
\description{
Query page for button and returns the first button that is succesfully clicked
}
