% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_news_item.R
\name{retrieve_news_items}
\alias{retrieve_news_items}
\title{Retrieve News Items}
\usage{
retrieve_news_items(
  news_outlet,
  topic,
  last_crawl = NULL,
  headless_browsing = NULL,
  handle_cookie_prompt = TRUE,
  verbose = FALSE,
  debug_output = FALSE
)
}
\arguments{
\item{news_outlet}{A string specifying the news outlet, either "dr" or "tv2".}

\item{topic}{A string specifying the topic to retrieve news from. Valid topics depend on the news outlet.}

\item{last_crawl}{A Date object or string specifying the date of the last crawl (optional).}

\item{headless_browsing}{Logical indicating if the browsing should be headless. Default is TRUE.}

\item{handle_cookie_prompt}{Logical indicating if the cookie prompt should be automatically handled. Default is TRUE.}

\item{verbose}{Logical indicating whether to print messages to the console.}

\item{debug_output}{Logical indicating whether to return additional object for debuging purposes}
}
\value{
A data frame of news items retrieved from the specified news outlet and topic.
}
\description{
Retrieves news items from specified news outlets and topics, with options for headless browsing and handling cookie prompts.
}
\examples{
\dontrun{
news_items <- retrieve_news_items("dr", "politik", Sys.Date() - 7)
news_items <- retrieve_news_items("tv2", "klima", last_crawl = "2023-01-01", headless_browsing = FALSE)
}
}
