% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapR_package.R
\docType{package}
\name{ScrapR}
\alias{ScrapR}
\title{Scrapr}
\description{
Scrapes news articles from dr.dk and tv2.dk.
The package has options to automatically handle cookie consent promts and
to press "view more" by using RSelenium to interact with the webpage.
}
\author{
Thomas Tesfaye-Berhane
}
