% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_to_seurat.R
\name{scraps_to_matrix}
\alias{scraps_to_matrix}
\title{Read scraps output from umi_tools to sparseMatrix}
\usage{
scraps_to_matrix(
  file,
  n_min = 5,
  gene_min = 5,
  alt_only = FALSE,
  filter_low = TRUE,
  cell_ids = NULL,
  types = "3'UTR",
  types2 = NULL,
  pf = NULL
)
}
\arguments{
\item{file}{scraps output table containing}

\item{n_min}{minimum number of observations for a site to be kept}

\item{gene_min}{minimum number of observations for gene to be kept}

\item{alt_only}{if TRUE, only keep genes with alternative polyA sites}

\item{filter_low}{if TRUE, remove low genes and cell ids, if FALSE, set to -1}

\item{cell_ids}{if given, use only these cell barcodes, and fill in empty ones}

\item{types}{filter to only these types of PA sites, set to NULL to use all}

\item{types2}{another filter to only these types of PA sites, set to NULL to use all}

\item{pf}{data.frame with SAF first column and position factor by gene, calculated by \code{parse_saf_pf}}
}
\value{
count matrix
}
\description{
Read scraps output from umi_tools to sparseMatrix
}
\examples{
s_small <- scraps_to_matrix("sample_R2_counts.tsv.gz",
                            alt_only = FALSE)
}
