% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_priming_region.R
\name{scraps_to_region_counts}
\alias{scraps_to_region_counts}
\title{Read scraps output to read counts by region type}
\usage{
scraps_to_region_counts(files, cell_ids = NULL)
}
\arguments{
\item{files}{scraps output table files containing}

\item{cell_ids}{if given, use only these cell barcodes, and fill in empty ones}
}
\value{
count list for regions
}
\description{
Read scraps output to read counts by region type
}
