% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraps_to_seurat.R
\name{scraps_to_seurat}
\alias{scraps_to_seurat}
\title{Add scraps output to existing seurat object as new assay}
\usage{
scraps_to_seurat(
  file,
  object,
  assay_name = "Asite",
  n_min = 5,
  alt_only = TRUE,
  types = "3'UTR",
  pf = NULL
)
}
\arguments{
\item{file}{scraps output table containing}

\item{object}{seurat object}

\item{n_min}{minimum number of observations for a site to be kept}

\item{alt_only}{if TRUE, only keep genes with alternative polyA sites}

\item{types}{filter to only these types of PA sites, set to NULL to use all}

\item{pf}{data.frame with SAF first column and position factor by gene, calculated by \code{parse_saf_pf}}
}
\value{
Seurat object with inserted assay
}
\description{
Add scraps output to existing seurat object as new assay
}
