% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simple_count_comparison}
\alias{simple_count_comparison}
\title{Simple Count Comparison}
\usage{
simple_count_comparison(
  count_table = NULL,
  target_date = NULL,
  count_tolerance = 1,
  time_window = 3
)
}
\arguments{
\item{count_table}{The current date-count table.}

\item{target_date}{The specific historical date of interest (as a Date object).}

\item{count_tolerance}{Acceptable difference in counts for considering them similar.}

\item{time_window}{Days around the current date to include in the comparison.}
}
\value{
TRUE if the count on the target date is similar to counts on recent dates, FALSE otherwise.
}
\description{
Compares the count of events on a specific historical date with counts on recent dates.
The function examines if the event frequency on a past date aligns closely with the event frequency on dates nearer to the present.
}
\examples{
\dontrun{
date_counts <- c("2017-12-24" = 2, "2022-12-24" = 2, "2023-10-25" = 1, "2023-10-31" = 1,
                 "2023-11-10" = 1, "2023-11-19" = 1, "2023-12-20" = 6, "2023-12-21" = 8,
                 "2023-12-22" = 10, "2023-12-23" = 9, "2023-12-24" = 14)
target_date <- as.Date("2023-11-10")
simple_count_comparison(date_counts, target_date)
}
}
