% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_PDF_data.R
\name{simulate_PDF_data}
\alias{simulate_PDF_data}
\title{Simulate data}
\usage{
simulate_PDF_data()
}
\description{
This function simulates data
}
\examples{
simulate_PDF_data()
}
\keyword{simulation}
