% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\title{Retrieve motorcycle prices}
\usage{
zoom(type)
}
\arguments{
\item{type}{right now six types--"2018-honda-xr650l", "2014-husaberg-fe-501", 
"2018-aprilia-rsv4-rf", "2018-beta-evo-250-factory", 
"2018-bmw-r-1200-gs-adventure"--are hard coded into \code{scrapr::collect}}
}
\value{
A data frame of one rows (types) and 4 columns: 
type of motocycle \code{cycle}, \code{price},
the url \code{site}, and \code{date}.
}
\description{
\code{zoom} works with \code{collect} to retrieve motorcycle prices
}
