% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{1036 x 8: Wal-API, query by Name results}
\format{An object of class \code{data.frame} with 1036 rows and 8 columns.}
\usage{
dat
}
\description{
1036 x 8: Wal-API, query by Name results
}
\keyword{datasets}
