% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_itemIDs}
\alias{dat_itemIDs}
\title{989 Walmart Item IDs}
\format{An object of class \code{integer} of length 989.}
\usage{
dat_itemIDs
}
\description{
989 Walmart Item IDs
}
\keyword{datasets}
