% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datr}
\alias{datr}
\title{989 x 8: Wal-API, query by ID results}
\format{An object of class \code{data.frame} with 989 rows and 8 columns.}
\usage{
datr
}
\description{
989 x 8: Wal-API, query by ID results
}
\keyword{datasets}
