% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_PDF_data.R
\name{load_PDF_data}
\alias{load_PDF_data}
\title{Load PDF data from figure}
\usage{
load_PDF_data(file_name = "figure1.pdf")
}
\arguments{
\item{file_name}{File to load}
}
\description{
This function extracts PDF data and creates a guide figure to align extraction
}
\examples{
load_PDF_data()
}
